+++ biomodels_qc/biomodels_qc/convert.py
@@ -7,15 +7,14 @@ such as BioPAX, MATLAB/Octave, and XPP.
 :License: MIT
 """
 
-from .utils import get_smbl_files_for_entry, are_biopax_files_the_same
-from .validation import validate_sbml_file, validate_xpp_file
+from .utils import get_smbl_files_for_entry, are_biopax_files_the_same, does_sbml_file_represent_core_kinetic_model
+from .validation import validate_sbml_file, validate_xpp_file, validate_octave_file
 from biosimulators_utils.sedml.data_model import Task
 from biosimulators_utils.sedml.io import SedmlSimulationReader
 from kisao import Kisao
 from kisao.utils import get_ode_algorithms
 import enum
 import glob
-import libsbml
 import os
 import shutil
 import subprocess
@@ -67,30 +66,25 @@ def convert_entry(dirname, alt_sbml_formats=None):
             pass
 
     for filename in get_smbl_files_for_entry(dirname, include_urn_files=False):
-        doc = libsbml.readSBMLFromFile(filename)
-        for i_plugin in range(doc.getNumPlugins()):
-            plugin = doc.getPlugin(i_plugin)
-            package_name = plugin.getPackageName()
-            if (
-                (has_sedml_task and not ode_simulation)
-                or package_name in ['arrays', 'comp', 'distrib', 'dyn', 'fbc', 'groups', 'math', 'multi', 'qual', 'spatial']
-            ):
-                for alt_sbml_format in [AltSbmlFormat.Matlab, AltSbmlFormat.Octave, AltSbmlFormat.XPP]:
-                    alt_sbml_formats.remove(alt_sbml_format)
+        if (
+            (has_sedml_task and not ode_simulation)
+            or not does_sbml_file_represent_core_kinetic_model(filename)
+        ):
+            for alt_sbml_format in [AltSbmlFormat.Matlab, AltSbmlFormat.Octave, AltSbmlFormat.XPP]:
+                alt_sbml_formats.remove(alt_sbml_format)
 
-                    if filename.endswith('_url.xml'):
-                        old_final_converted_filename = filename[0:-8] + ALT_SBML_FORMAT_DATA[alt_sbml_format]['old_final_extension']
-                        final_converted_filename = filename[0:-8] + ALT_SBML_FORMAT_DATA[alt_sbml_format]['final_extension']
-                    else:
-                        old_final_converted_filename = os.path.splitext(
-                            filename)[0] + ALT_SBML_FORMAT_DATA[alt_sbml_format]['old_final_extension']
-                        final_converted_filename = os.path.splitext(filename)[0] + ALT_SBML_FORMAT_DATA[alt_sbml_format]['final_extension']
+                if filename.endswith('_url.xml'):
+                    old_final_converted_filename = filename[0:-8] + ALT_SBML_FORMAT_DATA[alt_sbml_format]['old_final_extension']
+                    final_converted_filename = filename[0:-8] + ALT_SBML_FORMAT_DATA[alt_sbml_format]['final_extension']
+                else:
+                    old_final_converted_filename = os.path.splitext(
+                        filename)[0] + ALT_SBML_FORMAT_DATA[alt_sbml_format]['old_final_extension']
+                    final_converted_filename = os.path.splitext(filename)[0] + ALT_SBML_FORMAT_DATA[alt_sbml_format]['final_extension']
 
-                    if os.path.isfile(old_final_converted_filename):
-                        os.remove(old_final_converted_filename)
-                    if os.path.isfile(final_converted_filename):
-                        os.remove(final_converted_filename)
-                break
+                if os.path.isfile(old_final_converted_filename):
+                    os.remove(old_final_converted_filename)
+                if os.path.isfile(final_converted_filename):
+                    os.remove(final_converted_filename)
 
         for alt_sbml_format in alt_sbml_formats:
             try:
@@ -111,7 +105,11 @@ def convert_entry(dirname, alt_sbml_formats=None):
                 if move_to_alt_filename:
                     shutil.move(temp_filename, alt_filename)
 
-                if alt_sbml_format == AltSbmlFormat.SBML_URN and validate_sbml_file(alt_filename)[0]:
+                if alt_sbml_format == AltSbmlFormat.Octave and validate_octave_file(alt_filename)[0]:
+                    warnings.warn(termcolor.colored('`{}` could not be converted to valid Octave file.'.format(filename)), UserWarning)
+                    os.remove(alt_filename)
+
+                elif alt_sbml_format == AltSbmlFormat.SBML_URN and validate_sbml_file(alt_filename)[0]:
                     warnings.warn(termcolor.colored('`{}` could not be converted to valid SBML file.'.format(filename)), UserWarning)
                     os.remove(alt_filename)