+++ biomodels_qc/biomodels_qc/validation.py
@@ -46,10 +46,12 @@ __all__ = [
     'validate_image_file',
     'validate_ipynb_notebook_file',
     'validate_matlab_data_file',
+    'validate_octave_file',
     'validate_owl_ontology_file',
     'validate_pdf_file',
     'validate_python_file',
     'validate_sbml_file',
+    'validate_scilab_file',
     'validate_sedml_file',
     'validate_vcml_file',
     'validate_xml_file',
@@ -236,6 +238,30 @@ def validate_matlab_data_file(filename):
         return [[str(exception)]], []
 
 
+def validate_octave_file(filename):
+    """ Determine if an Octave file is valid
+
+    Args:
+        filename (:obj:`str`): path to Octave file
+
+    Returns:
+        :obj:`tuple`:
+
+            * nested :obj:`list` of :obj:`str`: nested list of errors
+            * nested :obj:`list` of :obj:`str`: nested list of warnings
+    """
+    result = subprocess.run(['octave', '--no-gui', '--no-window-system', '--quiet', filename],
+                            check=False, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    if result.returncode == 0:
+        stderr = result.stderr.decode(errors="ignore")
+        if stderr.startswith('parse error'):
+            return [['`{}` is not valid.'.format(filename), [[stderr]]]], []
+        else:
+            return [], []
+    else:
+        raise RuntimeError('Octave failed')
+
+
 def validate_owl_ontology_file(filename):
     """ Determine if an OWL ontology file is valid
 
@@ -321,6 +347,26 @@ def validate_sbml_file(filename):
             errors.append([sbml_error.getMessage()])
 
     return errors, warnings
+
+
+def validate_scilab_file(filename):
+    """ Determine if an Scilab file is valid
+
+    Args:
+        filename (:obj:`str`): path to Scilab file
+
+    Returns:
+        :obj:`tuple`:
+
+            * nested :obj:`list` of :obj:`str`: nested list of errors
+            * nested :obj:`list` of :obj:`str`: nested list of warnings
+    """
+    result = subprocess.run(['scilab', '--parse-file', filename, '-nb', '-nwni'],
+                            check=False, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    if result.returncode == 0:
+        return [], []
+    else:
+        return [['`{}` is not valid.'.format(filename)]], []
 
 
 def validate_sedml_file(filename, dirname=None, max_number_of_time_course_steps=1000, simulators=None):
@@ -433,7 +479,13 @@ def validate_vcml_file(filename):
             * nested :obj:`list` of :obj:`str`: nested list of errors
             * nested :obj:`list` of :obj:`str`: nested list of warnings
     """
-    return validate_xml_file(filename)
+    errors, warnings = validate_xml_file(filename)
+
+    with open(filename, 'rb') as file:
+        if 'SBML Model could not be automatically converted to VCML'.encode() in file.read():
+            errors.append(['`{}` does not represent a valid model.'.format(filename)])
+
+    return errors, warnings
 
 
 def validate_xml_file(filename):
@@ -576,6 +628,10 @@ EXTENSION_VALIDATOR_MAP = {
         'description': 'JPEG image',
         'validator': lambda filename: validate_image_file(filename, ImageFormat.jpeg),
     },
+    # '.m': {
+    #    'description': 'Octave script',
+    #    'validator': validate_octave_file,
+    # },
     '.mat': {
         'description': 'MATLAB data file',
         'validator': validate_matlab_data_file,
@@ -603,6 +659,10 @@ EXTENSION_VALIDATOR_MAP = {
     '.sbml': {
         'description': 'SBML',
         'validator': validate_sbml_file,
+    },
+    '.sci': {
+        'description': 'Scilab',
+        'validator': validate_scilab_file,
     },
     '.sedml': {
         'description': 'SED-ML',