+++ fix_manual_corrections.py
@@ -39,7 +39,7 @@ def run(id, working_dir):
             with open(filename, 'rb') as file:
                 contents = file.read()
 
-            assert fix['old'].encode() in contents, "Text could not be replaced in `{}`".format(filename)
+            assert fix['old'].encode() in contents, "Text could not be replaced in `{}`\n\n{}".format(filename, fix['old'])
             contents = contents.replace(fix['old'].encode(), fix['new'].encode())
 
             with open(filename, 'wb') as file: